WIP: Moving main() to a separate file from JogROSInterface.
