+++ moveit_experimental/jog_arm/CMakeLists.txt
@@ -1,8 +1,14 @@
-add_executable(jog_arm_server src/jog_arm/jog_arm_server.cpp)
+add_executable(jog_arm_server 
+  src/jog_arm/jog_arm_server.cpp
+  src/jog_arm/jog_ros_interface.cpp
+)
 add_dependencies(jog_arm_server ${catkin_EXPORTED_TARGETS})
 target_link_libraries(jog_arm_server ${catkin_LIBRARIES} ${Eigen_LIBRARIES})
 
-add_executable(spacenav_to_twist src/jog_arm/teleop_examples/spacenav_to_twist.cpp)
+add_executable(spacenav_to_twist 
+  src/jog_arm/teleop_examples/spacenav_to_twist.cpp
+  src/jog_arm/jog_ros_interface.cpp
+)
 add_dependencies(spacenav_to_twist ${catkin_EXPORTED_TARGETS})
 target_link_libraries(spacenav_to_twist ${catkin_LIBRARIES} ${Eigen_LIBRARIES})
